Decided to postpone Internalization, restructured site template
