+++ package.json
@@ -17,15 +17,12 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@heroicons/react": "^2.1.1",
     "framer-motion": "^11.0.3",
-    "i18next": "^23.8.2",
-    "i18next-browser-languagedetector": "^7.2.0",
-    "i18next-http-backend": "^2.4.3",
     "next": "14.1.0",
     "react": "^18",
     "react-dom": "^18",
-    "react-i18next": "^14.0.2",
     "react-icons": "^5.0.1",
     "react-type-animation": "^3.2.0",
+    "resend": "^3.1.0",
     "styled-components": "^6.1.8"
   },
   "devDependencies": {
