--- pages/LocaleContext.js
@@ -1,8 +1,0 @@
-import { createContext } from "react";
-
-const defaultValue = {
-    locale: 'en',
-    setLocale: () => {}
-}
-
-export default createContext(defaultValue);
