+++ pages/_app.js
@@ -1,6 +1,4 @@
 import "@/styles/globals.css";
-import i18n  from "./i18n";
-
 export default function App({ Component, pageProps }) {
   return <Component {...pageProps} />;
 }
