+++ pages/components/MenuOverlay.jsx
@@ -1,10 +1,10 @@
 import React from "react";
-import NavLink from "./NavLink";
+import NavLink from "../../outside-components/NavLink";
 
 const MenuOverlay = ({ links }) => {
   return (
     <ul className="flex flex-col py-4 items-center">
-      {links.map((link, index) => (
+      {links && links.map((link, index) => (
         <li key={index}>
           <NavLink href={link.path} title={link.title} />
         </li>
@@ -12,5 +12,16 @@ const MenuOverlay = ({ links }) => {
     </ul>
   );
 };
+  
+//   return (
+//     <ul className="flex flex-col py-4 items-center">
+//       {links.map((link, index) => (
+//         <li key={index}>
+//           <NavLink href={link.path} title={link.title} />
+//         </li>
+//       ))}
+//     </ul>
+//   );
+// };
 
 export default MenuOverlay;
