+++ pages/contact.js
@@ -51,7 +51,6 @@ width: 100%;
     align-items: flex-start;
     flex-direction: column;
     width: 100%;
-    height: 50%;
     font-size: 16px;
 
     input,
